♻️ Move DriverOptions and DefaultDriverOptions to opts
